mysite/pyproject.toml
[tool.poetry]
name = "factory-boy-best-practices"
version = "0.1.0"
description = ""
authors = ["Camila Maia <cmaiacd@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
pytest-django = "^4.4.0"
black = "21.6b0"
pre-commit = "^2.13.0"
ipdb = "^0.13.9"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
